move callback to functionEditor.js
