+++ apps/src/blockly/addons/functionEditor.js
@@ -27,3 +27,48 @@ export default class FunctionEditor {
   // TODO
   openWithNewFunction() {}
 }
+
+export function procedureFlyoutCallback(workspace) {
+  const xmlList = [];
+  if (Blockly.Blocks['procedures_defnoreturn']) {
+    // <block type="procedures_defnoreturn" gap="16">
+    //     <field name="NAME">do something</field>
+    // </block>
+    const block = Blockly.utils.xml.createElement('block');
+    block.setAttribute('type', 'procedures_defnoreturn');
+    block.setAttribute('gap', 16);
+    const nameField = Blockly.utils.xml.createElement('field');
+    nameField.setAttribute('name', 'NAME');
+    nameField.appendChild(
+      Blockly.utils.xml.createTextNode(
+        Blockly.Msg['PROCEDURES_DEFNORETURN_PROCEDURE']
+      )
+    );
+    block.appendChild(nameField);
+    xmlList.push(block);
+
+    if (xmlList.length) {
+      // Add slightly larger gap between system blocks and user calls.
+      xmlList[xmlList.length - 1].setAttribute('gap', 24);
+    }
+
+    const allProcedureCalls = Blockly.Procedures.allProcedures(workspace)[0];
+    for (let i = 0; i < allProcedureCalls.length; i++) {
+      const name = allProcedureCalls[i][0];
+      const args = allProcedureCalls[i][1];
+      const block = Blockly.utils.xml.createElement('block');
+      block.setAttribute('type', 'procedures_callnoreturn');
+      block.setAttribute('gap', 16);
+      const mutation = Blockly.utils.xml.createElement('mutation');
+      mutation.setAttribute('name', name);
+      block.appendChild(mutation);
+      for (let j = 0; j < args.length; j++) {
+        const arg = Blockly.utils.xml.createElement('arg');
+        arg.setAttribute('name', args[j]);
+        mutation.appendChild(arg);
+      }
+      xmlList.push(block);
+    }
+    return xmlList;
+  }
+}
